requirements.txt
asgiref==3.8.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
Django==5.1.7
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-storages==1.14.5
django-taggit==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
